requirements.txt
anyascii==0.1.7
bcrypt==3.2.0
beautifulsoup4==4.6.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==2.2.18
django-filter==2.4.0
django-heroku==0.3.1
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
Fabric3==1.14.post1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
l18n==2020.6.1
openpyxl==3.0.7
paramiko==2.7.2
Pillow==7.2.0
psycopg2==2.8.4
pycparser==2.20
PyNaCl==1.4.0
python-decouple==3.3
pytz==2020.1
requests==2.22.0
sentry-sdk==1.0.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.3.0
tablib==3.0.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7.4
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
